[SPARK-42052][SQL] Codegen Support for HiveSimpleUDF
